pyproject.toml
[project]
name = "master-summarizer"
version = "0.1.0"
description = "Multi-agent document summarization pipeline for POT Uribia, Guajira"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-docx>=1.1",
    "PyMuPDF>=1.24",
    "Pillow>=10.0",
    "chromadb>=0.5",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
summarizer = "src.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
